Update zoom extension
- Merge branch \'contributions/merge-1782464030084\'
- Pull contributions
- docs(readme): add comprehensive usage guide and feature list
- chore(scripts): update build and prepublishOnly scripts - Remove environment flag from build command - Add prepublishOnly script to prevent accidental npm publishes
- refactor(api): enhance Zoom API error handling and retry logic - Introduce retry mechanism for meeting list requests with exponential backoff - Simplify and modularize fetch functions for cleaner code structure - Update dependencies and contributors in package.json - Remove unused eslint configuration file - Expand .gitignore to exclude additional build and environment-specific files
