+++ extensions/zoom/.gitignore
@@ -12,5 +12,3 @@ compiled_raycast_rust
 
 # misc
 .DS_Store
-
-raycast-env.d.ts
