+++ extensions/zoom/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Zoom Changelog
 
+## [Improvements] - {PR_MERGE_DATE}
+
+- Retry meeting list requests when Zoom rate limits the API
+- Improve reliability when loading upcoming meetings concurrently
+
 ## [Security Fix] - 2026-03-17
 
 - Bump lodash/lodash-es to fix prototype pollution vulnerability (CVE-2025-13465)
