--- extensions/zoom/eslint.config.js
@@ -1,6 +1,0 @@
-const { defineConfig } = require("eslint/config");
-const raycastConfig = require("@raycast/eslint-config");
-
-module.exports = defineConfig([
-  ...raycastConfig,
-]);
