+++ extensions/zoom/package.json
@@ -9,7 +9,8 @@
   "contributors": [
     "Visual-Studio-Coder",
     "samuelkraft",
-    "jacperegrinedavies"
+    "jacperegrinedavies",
+    "0xdhrv"
   ],
   "platforms": [
     "macOS",
@@ -307,26 +308,26 @@
     ]
   },
   "dependencies": {
-    "@raycast/api": "^1.104.1",
-    "@raycast/utils": "^2.2.2",
-    "date-fns": "^4.1.0",
-    "lodash": "^4.17.23",
-    "node-fetch": "^3.3.2"
+    "@raycast/api": "^1.104.20",
+    "@raycast/utils": "^2.2.7",
+    "date-fns": "^4.4.0",
+    "lodash": "^4.18.1"
   },
   "devDependencies": {
     "@raycast/eslint-config": "^2.1.1",
-    "@types/lodash": "^4.17.0",
-    "@types/node": "^25.0.3",
-    "@types/react": "^19.2.7",
-    "eslint": "^9.39.2",
-    "prettier": "^3.7.4",
-    "typescript": "^5.9.3"
+    "@types/lodash": "^4.17.24",
+    "@types/node": "^26.0.1",
+    "@types/react": "^19.2.17",
+    "eslint": "^10.5.0",
+    "prettier": "^3.8.4",
+    "typescript": "^6.0.3"
   },
   "scripts": {
-    "build": "ray build -e dist",
+    "build": "ray build",
     "dev": "ray develop",
     "fix-lint": "ray lint --fix",
     "lint": "ray lint",
+    "prepublishOnly": "echo \"\\n\\nIt seems like you are trying to publish the Raycast extension to npm.\\n\\nIf you did intend to publish it to npm, remove the \\`prepublishOnly\\` script and rerun \\`npm publish\\` again.\\nIf you wanted to publish it to the Raycast Store instead, use \\`npm run publish\\` instead.\\n\\n\" && exit 1",
     "publish": "npx @raycast/api@latest publish"
   }
 }
