+++ extensions/zoom/tsconfig.json
@@ -1,14 +1,10 @@
 {
   "$schema": "https://json.schemastore.org/tsconfig",
   "display": "Node 16",
-  "include": [
-    "src/**/*",
-    "raycast-env.d.ts"
-  ],
+  "include": ["src/**/*", "raycast-env.d.ts"],
   "compilerOptions": {
-    "lib": [
-      "es2021"
-    ],
+    "lib": ["es2021"],
+    "types": ["node"],
     "module": "commonjs",
     "target": "es2021",
     "strict": true,
